% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dataset.R, R/Dataspace.R
\docType{class}
\name{DataSet}
\alias{DataSet}
\alias{DataSet-class}
\alias{c,DataSet-method}
\alias{cbind2,DataSet,matrix-method}
\alias{checkParamBoundaries}
\alias{checkParamBoundaries,DataSet,numeric,numeric-method}
\alias{closeh5,DataSet-method}
\alias{closeh5,DataSpace-method}
\alias{extendDataSet}
\alias{extendDataSet,DataSet,numeric-method}
\alias{rbind2,DataSet,matrix-method}
\alias{readDataSet}
\alias{readDataSet,DataSet-method}
\alias{selectDataSpace}
\alias{selectDataSpace,DataSet-method}
\alias{writeDataSet}
\alias{writeDataSet,DataSet-method}
\title{The DataSet class}
\usage{
checkParamBoundaries(.Object, offset, count)

\S4method{checkParamBoundaries}{DataSet,numeric,numeric}(.Object, offset, count)

selectDataSpace(.Object, offset = rep(NA_integer_, length(.Object@dim)),
  count = rep(NA_integer_, length(.Object@dim)))

\S4method{selectDataSpace}{DataSet}(.Object, offset = rep(NA_integer_,
  length(.Object@dim)), count = rep(NA_integer_, length(.Object@dim)))

writeDataSet(.Object, data, dspace = selectDataSpace(.Object, rep(NA_integer_,
  length(.Object@dim)), GetDimensions(data)))

\S4method{writeDataSet}{DataSet}(.Object, data,
  dspace = selectDataSpace(.Object, rep(NA_integer_, length(.Object@dim)),
  GetDimensions(data)))

readDataSet(.Object, dspace = selectDataSpace(.Object, offset =
  rep(NA_integer_, length(.Object@dim)), count = rep(NA_integer_,
  length(.Object@dim))))

\S4method{readDataSet}{DataSet}(.Object, dspace = selectDataSpace(.Object,
  offset = rep(NA_integer_, length(.Object@dim)), count = rep(NA_integer_,
  length(.Object@dim))))

\S4method{closeh5}{DataSet}(.Object)

extendDataSet(.Object, dims)

\S4method{extendDataSet}{DataSet,numeric}(.Object, dims)

\S4method{rbind2}{DataSet,matrix}(x, y)

\S4method{cbind2}{DataSet,matrix}(x, y)

\S4method{c}{DataSet}(x, ..., recursive = FALSE)

\S4method{closeh5}{DataSpace}(.Object)
}
\arguments{
\item{.Object,x}{DataSet; S4 object of class \code{DataSet};}

\item{offset}{numeric; Offset to be selected from Hyperslab.}

\item{count}{numeric; Count to be selected from Hyperslab.}

\item{data,y}{object; Object to be stored in HDF5 file, can be either of type vector, matrix or array.}

\item{dspace}{DataSpace; Data space object used for data selection.}

\item{dims}{numeric; Dimensions of DataSet.}

\item{...}{additional arguments passed to \code{\link{c}}.}

\item{recursive}{logical; Argument passed to \code{\link{c}}.}
}
\description{
The DataSet class used to store data objects in HDF5 tree.
}

