% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dataspace.R
\docType{class}
\name{DataSpace}
\alias{DataSpace}
\alias{DataSpace-class}
\title{The DataSpace class}
\arguments{
\item{.Object}{DataSpace; S4 object of class \code{DataSpace};}

\item{offset}{numeric; Offset to be selected from Hyperslab.}

\item{count}{numeric; Count to be selected from Hyperslab.}

\item{dims}{numeric; Dimensions of DataSpace.}

\item{recursive}{logical; Argument passed to \code{\link{c}}.}

\item{...}{additional arguments passed to \code{\link{c}}.}
}
\description{
The DataSpace class used to store data objects in HDF5 tree.
}

