% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFG.R
\docType{methods}
\name{h5-deprecated}
\alias{h5-deprecated}
\alias{[<-,CommonFG,character,ANY-method}
\title{Deprecated CommonFG methods}
\usage{
\S4method{[}{CommonFG,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{CommonFG,character,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x, i, j, drop, value, ...}{See \code{\link{CommonFG}}}
}
\description{
Since \code{obj[name]} is commonly used to get and set multiple elements,
\code{obj[[name]]} is the better fit for setting or retrieving a single dataset or groups.
}
