% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H5File}
\alias{H5File}
\alias{H5File-class}
\alias{closeh5,H5File-method}
\title{The H5File class}
\usage{
\S4method{closeh5}{H5File}(.Object)
}
\arguments{
\item{.Object}{H5File; S4 object of class \code{H5File};}
}
\description{
Class to represent an HDF5 file.
}

