% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H5Group}
\alias{H5Group}
\alias{H5Group-class}
\alias{closeh5,H5Group-method}
\title{The H5Group class}
\usage{
\S4method{closeh5}{H5Group}(.Object)
}
\arguments{
\item{.Object}{H5Group; S4 object of class \code{H5Group};}
}
\description{
Class to represent an HDF5 Group.
}

