% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/H5Location.R
\docType{class}
\name{H5Location}
\alias{H5Location}
\alias{H5Location-class}
\title{The H5Location class}
\arguments{
\item{.Object}{H5Location; S4 object of class \code{H5Location};}

\item{attributename}{character; Name of attribute to be read/created.}

\item{data}{object; Data object to be used for attribute creation,
can be either of type vector, matrix or array.}
}
\description{
This class contains an example. This line goes into the description
}
\details{
This line and the next ones go into the details.
This line thus appears in the details as well.
}

