% Generated by roxygen2 (4.0.1): do not edit by hand
\name{h5ReadMemDF}
\alias{h5ReadMemDF}
\title{Function to read a data frame that has been stored in memory when given the external pointer}
\usage{
h5ReadMemDF(ptr)
}
\arguments{
\item{ptr}{externalptr to data frame stored in memory}
}
\value{
the data frame that is held at the externalptr
}
\description{
Function to read a data frame that has been stored in memory when given the external pointer
}

