% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dataset.R, R/Dataspace.R
\docType{class}
\name{DataSet}
\alias{DataSet}
\alias{DataSet-class}
\alias{[,DataSet,ANY,ANY,ANY-method}
\alias{[,DataSet,missing,missing,ANY-method}
\alias{[,DataSet,missing,numeric,ANY-method}
\alias{[,DataSet,numeric,missing,ANY-method}
\alias{[<-,DataSet,ANY,ANY-method}
\alias{[<-,DataSet,missing,missing-method}
\alias{[<-,DataSet,missing,numeric-method}
\alias{[<-,DataSet,numeric,missing-method}
\alias{c,DataSet-method}
\alias{cbind.DataSet}
\alias{cbind2,DataSet,matrix-method}
\alias{checkParamBoundaries}
\alias{checkParamBoundaries,DataSet,numeric,numeric-method}
\alias{closeh5,DataSet-method}
\alias{closeh5,DataSpace-method}
\alias{extendDataSet}
\alias{extendDataSet,DataSet,numeric-method}
\alias{rbind.DataSet}
\alias{rbind2,DataSet,matrix-method}
\alias{readDataSet}
\alias{readDataSet,DataSet-method}
\alias{selectDataSpace}
\alias{selectDataSpace,DataSet,ANY,ANY,missing-method}
\alias{selectDataSpace,DataSet,missing,missing,matrix-method}
\alias{selectDataSpace,DataSet,missing,missing,missing-method}
\alias{writeDataSet}
\alias{writeDataSet,DataSet-method}
\title{The DataSet class}
\usage{
checkParamBoundaries(.Object, offset, count)

\S4method{checkParamBoundaries}{DataSet,numeric,numeric}(.Object, offset, count)

selectDataSpace(.Object, offset = rep(NA_integer_, length(.Object@dim)),
  count = rep(NA_integer_, length(.Object@dim)), elem)

\S4method{selectDataSpace}{DataSet,missing,missing,missing}(.Object)

\S4method{selectDataSpace}{DataSet,ANY,ANY,missing}(.Object, offset, count)

\S4method{selectDataSpace}{DataSet,missing,missing,matrix}(.Object, elem)

writeDataSet(.Object, data, dspace = selectDataSpace(.Object, rep(NA_integer_,
  length(.Object@dim)), GetDimensions(data)), transpose = TRUE)

\S4method{writeDataSet}{DataSet}(.Object, data,
  dspace = selectDataSpace(.Object, rep(NA_integer_, length(.Object@dim)),
  GetDimensions(data)), transpose = TRUE)

readDataSet(.Object, dspace = selectDataSpace(.Object, offset =
  rep(NA_integer_, length(.Object@dim)), count = rep(NA_integer_,
  length(.Object@dim))))

\S4method{readDataSet}{DataSet}(.Object, dspace = selectDataSpace(.Object,
  offset = rep(NA_integer_, length(.Object@dim)), count = rep(NA_integer_,
  length(.Object@dim))))

\S4method{closeh5}{DataSet}(.Object)

extendDataSet(.Object, dims)

\S4method{extendDataSet}{DataSet,numeric}(.Object, dims)

rbind.DataSet(..., deparse.level = 1)

\S4method{rbind2}{DataSet,matrix}(x, y)

cbind.DataSet(..., deparse.level = 1)

\S4method{cbind2}{DataSet,matrix}(x, y)

\S4method{c}{DataSet}(x, ..., recursive = FALSE)

\S4method{[}{DataSet,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{DataSet,missing,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{DataSet,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{DataSet,missing,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{DataSet,missing,missing}(x, i, j, ...) <- value

\S4method{[}{DataSet,numeric,missing}(x, i, j, ...) <- value

\S4method{[}{DataSet,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{DataSet,ANY,ANY}(x, i, j, ...) <- value

\S4method{closeh5}{DataSpace}(.Object)
}
\arguments{
\item{.Object,x}{DataSet; S4 object of class \code{DataSet};}

\item{offset}{numeric; Offset to be selected from Hyperslab.}

\item{count}{numeric; Count to be selected from Hyperslab.}

\item{elem}{matrix; Matrix specifying element selection coordinates. Columns specify rank, rows specify different points.}

\item{data,y}{object; Object to be stored in HDF5 file, can be either of type vector, matrix or array.}

\item{dspace}{DataSpace; Data space object used for data selection.}

\item{transpose}{logical; Determine if data object (if is array) should be transposed.}

\item{dims}{numeric; Dimensions of DataSet.}

\item{...}{additional arguments passed to \code{\link{c}}.}

\item{deparse.level}{numeric; Not implemented yet.}

\item{recursive}{logical; Argument passed to \code{\link{c}}.}

\item{i}{integer; row index}

\item{j}{integer; column index}

\item{drop}{logical; specify if}

\item{value}{object; Value to be assigned to dataset}
}
\description{
The DataSet class used to store data objects in HDF5 tree.
}

