% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{class}
\name{DataSet}
\alias{DataSet}
\alias{DataSet-class}
\alias{closeh5,DataSet-method}
\alias{readDataSet}
\alias{readDataSet,DataSet-method}
\alias{writeDataSet}
\alias{writeDataSet,DataSet-method}
\title{The DataSet class}
\usage{
writeDataSet(.Object, data)

\S4method{writeDataSet}{DataSet}(.Object, data)

readDataSet(.Object)

\S4method{readDataSet}{DataSet}(.Object)

\S4method{closeh5}{DataSet}(.Object)
}
\arguments{
\item{.Object}{DataSet; S4 object of class \code{DataSet};}

\item{data}{object; Object to be stored in HDF5 file, can be either of type vector, matrix or array.}
}
\description{
The DataSet class used to store data objects in HDF5 tree.
}

